Added CORS configuration to allow frontend to access backend Also fixed the search functionality to work with the new CORS configuration This includes changes to the frontend and backend to allow for cross origin requests New features include: add stock update stock (done directly through the table) delete stock

+++ src/main/java/com/simpleStockInfo/demo/model/Stock.java
@@ -20,7 +20,7 @@ public class Stock {
     // Fields for the stock object same exact fields as the database table
     private String company_name;
     private double price;
-    private double market_cap;
+    private long market_cap;
     private double pe_ratio;
     private String sector;
 }
